refactor: Extract Bitcoin node management into dedicated module
This commit moves the Bitcoin node initialization and management code from main.rs to a new bitcoin_node module:
- Created bitcoin_node/mod.rs with BitcoinNode struct implementation
- Moved existing BitcoinNode code from main.rs to the new module
- Updated main.rs to import BitcoinNode from the new module
- Made minor adjustments to improve module visibility and organization
